AccountNo's appearing after set to int instead of Long datatype
accountno working but now to fix error relating to the debitamount

+++ src/main/java/com/groupg/onlinebanking/models/Transactions.java
@@ -56,7 +56,7 @@ public class Transactions {
     public void setDescription(String description) {
         this.description = description;
     }    
-    public long getAccountNo() {
+    public int getAccountNo() {
         return accountNo;
     }
